+++ CHANGES.md
@@ -3,10 +3,12 @@
 ### Major changes
 
 - Added the (optional) generation of a JSON-LD `@context` file
+- Added, in the YAML format, the possibility for a `dataset` entry on properties, resulting in the `"@container":"@graph"` statement in the generated context
 
 ### Minor changes, bug fixes
 
 - Better use of the Commander package for the CLI
+- Some CSV leftover still persisted in the README.md file...
 
 ## Version 1.1